requirements.txt
pillow
stepic
tkinterdnd2
